+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/AccountAddress.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.transactions;
 
+import com.concordium.sdk.crypto.SHA256;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
@@ -42,6 +43,15 @@ public final class AccountAddress {
     public static AccountAddress from(String address) {
         val addressBytes = Base58.decodeChecked(VERSION, address);
         return AccountAddress.from(addressBytes);
+    }
+
+    /**
+     * Create a {@link AccountAddress} from a {@link CredentialRegistrationId}.
+     * @param regId the credential registration id to derive the {@link AccountAddress} from
+     * @return the derived {@link AccountAddress}
+     */
+    public static AccountAddress from(CredentialRegistrationId regId) {
+        return AccountAddress.from(SHA256.hash(regId.getRegId()));
     }
 
     /**